add controller for like, relocate comment view and autofix rubocop
