+++ app/controllers/comments_controller.rb
@@ -10,10 +10,12 @@ class CommentsController < ApplicationController
       render 'posts/show'
     end
   end
+
   def new
     @post = Post.find(params[:post_id])
     @comment = Comment.new
   end
+
   private
 
   def comment_params
